% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colleagues.R
\name{get_player_colleagues}
\alias{get_player_colleagues}
\title{Get A Single Player's Colleagues}
\usage{
get_player_colleagues(all_players, player)
}
\arguments{
\item{all_players}{Data.frame. Data fetched with \code{\link{get_players}}.}

\item{player}{Character. A single player's name (case sensitive).}
}
\value{
A data.frame/tibble.
}
\description{
Get A Single Player's Colleagues
}
\examples{
\dontrun{get_player_colleagues()}
}
