% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players.R
\name{get_players}
\alias{get_players}
\title{Fetch A Dataframe From Transfermarkt}
\usage{
get_players(years = 1992:2021, country = "England")
}
\arguments{
\item{years}{Numeric. Years for which to fetch data.}

\item{country}{Character. Country from which to fetch data.}
}
\value{
A data.frame/tibble.
}
\description{
Fetch A Dataframe From Transfermarkt
}
\examples{
\dontrun{get_players()}
}
