% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{open_colleagues_quiz}
\alias{open_colleagues_quiz}
\title{Open 'Guess The Player' Quiz App}
\usage{
open_colleagues_quiz(all_players)
}
\arguments{
\item{all_players}{Data.frame. Data fetched from
\href{https://www.transfermarkt.com/}{Transfermarkt} with
\code{\link{get_players}}.}
}
\value{
Opens a Shiny app.
}
\description{
Open 'Guess The Player' Quiz App
}
\examples{
\dontrun{
# Fetch player data from Transfermarkt
epl_players <- get_players(1992:2020, "England")

# Open Shiny App
open_colleagues_quiz(epl_players)
}
}
