% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colleagues.R
\name{sample_co_colleagues}
\alias{sample_co_colleagues}
\title{Sample From Multiple Players' Co-Colleagues}
\usage{
sample_co_colleagues(all_players, players, n = 1)
}
\arguments{
\item{all_players}{Data.frame. Data fetched with \code{\link{get_players}}.}

\item{players}{Character. Two or more player names (case sensitive).}

\item{n}{Numeric. Number of team-mates' names to return.}
}
\value{
Character vector.
}
\description{
Sample From Multiple Players' Co-Colleagues
}
\examples{
\dontrun{sample_co_colleagues()}
}
