% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colleagues.R
\name{sample_colleagues}
\alias{sample_colleagues}
\title{Sample From Colleagues}
\usage{
sample_colleagues(all_players, players, n = NULL)
}
\arguments{
\item{all_players}{Data.frame. Data fetched with \code{\link{get_players}}.}

\item{players}{Character. Two or more player names (case sensitive).}

\item{n}{Numeric. Number of team-mates' names to return. If \code{NULL}, then
a default values of \code{5} will be used if the length of 'players' is 1,
and a default value of  \code{1} will be used if the length of 'players' is
two or more.}
}
\value{
Character vector.
}
\description{
Sample From Colleagues
}
\examples{
\dontrun{sample_colleagues()}
}
