% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colleagues.R
\name{sample_player_colleagues}
\alias{sample_player_colleagues}
\title{Sample From A Single Player's Colleagues}
\usage{
sample_player_colleagues(all_players, player, n = 5)
}
\arguments{
\item{all_players}{Data.frame. Data fetched with \code{\link{get_players}}.}

\item{n}{Numeric. Number of team-mates' names to return.}

\item{players}{Character. A single player name (case sensitive).}
}
\value{
Character vector.
}
\description{
Sample From A Single Player's Colleagues
}
\examples{
\dontrun{sample_player_colleagues()}
}
