% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccercolleagues-package.R
\docType{package}
\name{soccercolleagues-package}
\alias{soccercolleagues}
\alias{soccercolleagues-package}
\title{soccercolleagues: Find Common Team Mates For Soccer Players}
\description{
Identify a football player's current and former team mates. Find common team-mates for a set of named footballers. Data fetched from <transfermarkt.com>.
}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

}
\keyword{internal}
