% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colleagues.R
\name{get_colleagues}
\alias{get_colleagues}
\title{Get Player Colleagues}
\usage{
get_colleagues(all_players, players)
}
\arguments{
\item{all_players}{Data.frame. Data fetched with \code{\link{get_players}}.}

\item{players}{Character. Two or more player names (case sensitive).}
}
\value{
Character vector.
}
\description{
Get Player Colleagues
}
\examples{
\dontrun{get_co_colleagues()}
}
