% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{open_colleagues_quiz}
\alias{open_colleagues_quiz}
\title{Open 'Guess The Player' Quiz App}
\usage{
open_colleagues_quiz(all_players)
}
\arguments{
\item{all_players}{Data.frame. Data fetched with \code{\link{get_players}}.}
}
\value{
Opens a Shiny app.
}
\description{
Open 'Guess The Player' Quiz App
}
\examples{
\dontrun{open_colleagues_quiz()}
}
